Train shape parameters directly with mse loss.
Requires having the parameters in the first place for which a script is
added to fit a deformable model to the 68 keypoints which should then be
present in the dataset.

Moreover, add transition from parameter based loss to mse on keypoints
as epochs go by. This is inspired by the 3ddfa2 paper.

Do some cleanup too.